+++ src/core/draw/DrawPdf.ts
@@ -624,7 +624,9 @@ export class DrawPdf {
     if (positionContext.isTable) {
       const elementList = this.getOriginalElementList()
       const td = this.position.getTableTdByContext(elementList, positionContext)
-      const tdPadding = this.getTdPadding()
+      const {
+        table: { tdPadding }
+      } = this.options
       return td!.width! - tdPadding[1] - tdPadding[3]
     }
     // 分栏布局下按栏宽计算可用宽度（栏宽为缩放值，还原为未缩放单位）